src/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose
paramiko
python-multipart
python-dotenv
slowapi # rate limiting
celery
cryptography
redis